chore : apply MVP pattern of DetailView
