+++ teawon/project/toyproejct/app/src/main/java/com/example/toy_proejct/scenarios/detail/DetailViewModel.kt
@@ -1,0 +1,9 @@
+package com.example.toy_proejct.scenarios.detail
+
+import androidx.lifecycle.ViewModel
+
+
+class DetailViewModel : ViewModel() {
+
+
+}
