--- teawon/project/toyproejct/app/src/main/java/com/example/toy_proejct/scenarios/home/DetailActivity.kt
@@ -1,70 +1,0 @@
-package com.example.toy_proejct.scenarios.home
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
-import androidx.compose.material.Icon
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Surface
-import androidx.compose.material.Text
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowBack
-import androidx.compose.material.icons.filled.Favorite
-import androidx.compose.material.icons.filled.FavoriteBorder
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.example.toy_proejct.api.getSearchList.ProductListDto
-import com.example.toy_proejct.scenarios.home.ui.theme.ToyproejctTheme
-
-class DetailActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        val title = intent.getStringExtra("title")
-        val minimumPrice = intent.getIntExtra("minimumPrice", -1)
-
-        var isFavorite by mutableStateOf(false)
-
-        setContent {
-            Column(modifier = Modifier.fillMaxSize()) {
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(20.dp),
-                    horizontalArrangement = Arrangement.SpaceBetween,
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Icon(
-                        imageVector = Icons.Default.ArrowBack,
-                        contentDescription = "back",
-                        modifier = Modifier.clickable {
-                            finish()
-                        }
-                    )
-                    Text(
-                        text = title.toString(),
-                        modifier = Modifier.fillMaxWidth(),
-                        textAlign = TextAlign.Center
-                    )
-                    Icon(
-                        imageVector = if (isFavorite) Icons.Default.Favorite else Icons.Default.FavoriteBorder,
-                        contentDescription = "back",
-                        modifier = Modifier.clickable {
-                            isFavorite = !isFavorite
-                        }
-                    )
-                }
-                Text(text = "가격: ${minimumPrice}원")
-            }
-        }
-    }
-}
